Phase 7: Streamlit triage console (C1, C5)
    streamlit run app/streamlit_app.py

Reads the committed artefacts, so it starts cold with no pipeline run and shows
exactly the numbers in the README. No API keys, no secrets, no network calls --
a test asserts the source contains none of them (C5).

Four tabs: case queue, performance, entity drill-down, method. The queue is one
row per alerted ENTITY rather than per flagged transaction, matching the cost
model's unit -- an entity with six flagged rows is one investigation. Threshold
and analyst-capacity sliders update precision, recall, expected cost and
alerts-per-week live; the naive $10k comparison is a toggle.

The drill-down answers "why is this here" with a transaction timeline, the
reason codes as written by the rules engine, a feature profile of the
peak-risk transaction against the population, and a counterparty network graph
showing entities sharing counterparties with the subject. The graph shows
counterparties shared with 2-12 others rather than all of them, and says on the
page that employers and utilities are legitimately shared by hundreds, so a
dense picture is not evidence on its own.

MASKING IS STRUCTURAL, NOT INCIDENTAL. Every table goes through one show_table
helper that masks first, and a test parses the app's AST to assert show_table
is the only caller of st.dataframe anywhere in the file. Behavioural tests
catch a panel that builds the wrong frame; that source check catches the panel
somebody adds next year with no test of its own. A second source check catches
raw identifiers interpolated into captions and titles, where masking the table
would not have helped.

Masking covers who, not what: entity and counterparty identifiers and display
names are masked, case and transaction references are not, because those are
the references an analyst quotes when escalating.

Two real bugs found by running the thing rather than by reading it:

  * Streamlit parses `$...$` as inline LaTeX, so a reason code carrying two
    currency amounts lost both dollar signs and set the text between them as
    maths: "11 senders paid in 10,995.55 ... (avg 999.60)". Escaping happens at
    the render boundary, since the reason strings also go to CSV and JSON where
    they must stay clean. Pinned by a test that fails if any st.markdown or
    st.caption line passes an unescaped dollar sign.

  * The feature profile did arithmetic on a row sliced out of a mixed-dtype
    frame, which comes back as object dtype and silently downcasts. Surfaced
    because pyproject turns FutureWarning into an error.

Verified live: all four tabs render, health endpoint 200, screenshot in
docs/dashboard_case_queue.png showing masked identifiers (ENT-...1292) and
names reduced to initials.

154 tests passing.

+++ src/anomaly_model.py
@@ -226,7 +226,7 @@ def top_feature_deviations(features: pd.DataFrame, transaction_id: str,
     if row.empty:
         return pd.DataFrame(columns=["feature", "value", "population_mean", "deviation_sigma"])
 
-    deviation = ((row.iloc[0] - means) / stds).fillna(0.0)
+    deviation = ((row.iloc[0].astype(float) - means) / stds).fillna(0.0)
     ordered = deviation.abs().sort_values(ascending=False).head(top_n).index
     return pd.DataFrame(
         {
